Image Classification using CNN/requirements.txt
tensorflow>=2.0
h5py
matplotlib
opencv-python
numpy
streamlit
pillow
